improved context , add copy when code block
